+++ bindings/python/src/digests/mod.rs
@@ -1,40 +1,65 @@
 // This is intended to provide minimal Python bindings to functions in the `digests` module of the `gtars` crate.
 
+use gtars::digests::{md5, sha512t24u, DigestResult};
 use pyo3::prelude::*;
-use gtars::digests::{sha512t24u, md5, DigestResult};
+
+use pyo3::exceptions::PyTypeError;
+use pyo3::prelude::*;
+use pyo3::types::{PyBytes, PyString};
 
 #[pyfunction]
-pub fn sha512t24u_digest(readable: &str) -> String {
-    return sha512t24u(readable);
+pub fn sha512t24u_digest(readable: &Bound<'_, PyAny>) -> PyResult<String> {
+    if let Ok(s) = readable.downcast::<PyString>() {
+        Ok(sha512t24u(s.to_str()?)) // Borrowed, no copying
+    } else if let Ok(b) = readable.downcast::<PyBytes>() {
+        Ok(sha512t24u(b.as_bytes())) // Borrowed, no copying
+    } else {
+        Err(PyTypeError::new_err("Expected str or bytes"))
+    }
 }
 
 #[pyfunction]
-pub fn md5_digest(readable: &str) -> String {
-    return md5(readable);
+pub fn md5_digest(readable: &Bound<'_, PyAny>) -> PyResult<String> {
+    if let Ok(s) = readable.downcast::<PyString>() {
+        Ok(md5(s.to_str()?)) // Borrowed, no copying
+    } else if let Ok(b) = readable.downcast::<PyBytes>() {
+        Ok(md5(b.as_bytes())) // Borrowed, no copying
+    } else {
+        Err(PyTypeError::new_err("Expected str or bytes"))
+    }
 }
 
+// This can take either a PosixPath or a string
+// The `&Bound<'_, PyAny>` references any Python object, bound to the Python runtime.
 #[pyfunction]
-pub fn digest_fasta(fasta: &str) -> PyResult<Vec<PyDigestResult>> {
-    match gtars::digests::digest_fasta(fasta) {
+pub fn digest_fasta(fasta: &Bound<'_, PyAny>) -> PyResult<Vec<PyDigestResult>> {
+    let fasta = fasta.to_string();
+    match gtars::digests::digest_fasta(&fasta) {
         Ok(digest_results) => {
-            let py_digest_results: Vec<PyDigestResult> = digest_results.into_iter().map(PyDigestResult::from).collect();
+            let py_digest_results: Vec<PyDigestResult> = digest_results
+                .into_iter()
+                .map(PyDigestResult::from)
+                .collect();
             Ok(py_digest_results)
-        },
-        Err(e) => Err(PyErr::new::<pyo3::exceptions::PyIOError, _>(format!("Error processing FASTA file: {}", e))),
+        }
+        Err(e) => Err(PyErr::new::<pyo3::exceptions::PyIOError, _>(format!(
+            "Error processing FASTA file: {}",
+            e
+        ))),
     }
 }
 
 #[pyclass]
-#[pyo3(name="DigestResult")]
+#[pyo3(name = "DigestResult")]
 pub struct PyDigestResult {
-    #[pyo3(get,set)]
+    #[pyo3(get, set)]
     pub id: String,
-    #[pyo3(get,set)]
+    #[pyo3(get, set)]
     pub length: usize,
-    #[pyo3(get,set)]
+    #[pyo3(get, set)]
     pub sha512t24u: String,
-    #[pyo3(get,set)]
-    pub md5: String
+    #[pyo3(get, set)]
+    pub md5: String,
 }
 
 #[pymethods]
@@ -44,7 +69,10 @@ impl PyDigestResult {
     }
 
     fn __str__(&self) -> PyResult<String> {
-        Ok(format!("DigestResult for sequence {}\n  length: {}\n  sha512t24u: {}\n  md5: {}", self.id, self.length, self.sha512t24u, self.md5))
+        Ok(format!(
+            "DigestResult for sequence {}\n  length: {}\n  sha512t24u: {}\n  md5: {}",
+            self.id, self.length, self.sha512t24u, self.md5
+        ))
     }
 }
 
@@ -54,7 +82,7 @@ impl From<DigestResult> for PyDigestResult {
             id: value.id,
             length: value.length,
             sha512t24u: value.sha512t24u,
-            md5: value.md5
+            md5: value.md5,
         }
     }
 }
@@ -68,4 +96,3 @@ pub fn digests(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<PyDigestResult>()?;
     Ok(())
 }
-