+++ bindings/python/src/utils/mod.rs
@@ -20,7 +20,7 @@ pub fn extract_regions_from_py_any(regions: &Bound<'_, PyAny>) -> Result<RegionS
             .into());
         }
 
-        let regions = gtars::common::utils::extract_regions_from_bed_file(regions);
+        let regions = gtars::common::models::RegionSet::try_from(regions);
         match regions {
             Ok(regions) => return Ok(RegionSet::from(regions)),
             Err(e) => return Err(pyo3::exceptions::PyValueError::new_err(e.to_string()).into()),
@@ -44,7 +44,12 @@ pub fn extract_regions_from_py_any(regions: &Bound<'_, PyAny>) -> Result<RegionS
             let start = x.getattr("start").unwrap().extract::<u32>().unwrap();
             let end = x.getattr("end").unwrap().extract::<u32>().unwrap();
 
-            Ok(Region { chr, start, end })
+            Ok(Region {
+                chr,
+                start,
+                end,
+                rest: None,
+            })
         })
         .collect::<Vec<_>>();